refactor: moved NewClaim to containers
