+++ src/components/navigation/DashboardStack.tsx
@@ -1,7 +1,7 @@
 import { createStackNavigator } from 'react-navigation-stack'
 import { DASHBOARD, NEW_CLAIM, SEND_FOR_VERIFICATION } from '../../routes'
 import Dashboard from '../../containers/Dashboard'
-import NewClaim from '../NewClaim'
+import NewClaim from '../../containers/NewClaim'
 import SendForVerification from '../SendForVerification'
 
 const DashboardStack = createStackNavigator({
